done, just missing some tesing.
Testing of endpoints done. look in the test folder: excerciseWithJavalinAndCrud -> controllers -> RestAssurredHotelTest and RestAssuedRoomTests.
Had planned on making the tests for my DAO as well, but the security part of the assignment ended up taking all me time.

We had to continue on the project from last week, so you will find the assignment done in week 5. will give a description of where to find the individual tasks below:

Part 1 Hashing passwords - Completed: Look under UserController's login and register method. The actual hashing of the password is done in prepersist on the User Entity.

part 2 Login and Register Endpoints - Completed:
Both logic for both the Register and Login part can the found in routes and followed through the code.
Authenticate part added as a before to the protected routes in Routes.
Part 3 - Protect Endpoints - Completed: Look under getPotectedRoutes in Routes